add game states

+++ src/engine/engine.ts
@@ -1,8 +1,10 @@
 import * as PIXI from 'pixi.js';
-import { EntityManager, System, RenderableSystem } from './entity';
+import { EntityManager } from './entity';
+import { System, RenderableSystem } from './system';
 import { EventQueue } from './events';
 import { GameProperties } from './gameprops';
 import { GameState } from './state';
+import { StateTransitionDescription, StateTransitionSystem } from './statetransition';
 
 
 export class Engine {
@@ -17,7 +19,11 @@ export class Engine {
         transparent: false,
         resolution: window.devicePixelRatio
     };
+    private stateTransitions: StateTransitionDescription;
+    private stateTransitionSystem: StateTransitionSystem;
     private currentState: GameState;
+    private running: boolean;
+    private pendingActions: (()=>void)[];
 
     constructor(properties: GameProperties) {
         this.eventQueue = new EventQueue();
@@ -25,35 +31,74 @@ export class Engine {
         this.entityManager = new EntityManager(properties, this.eventQueue),
         this.pixiApp = new PIXI.Application({...this.pixiDefaultProperties, ...properties.pixiProperties});
         this.pixiApp.stage.sortableChildren = true;
+        this.running = false;
+        this.pendingActions = [];
     }
 
-    setState(state: GameState) {
-        this.pixiApp.ticker.remove( this.gameLoop, this );
-        if (this.currentState) {
-            this.currentState.destroySystems();
-            this.currentState.destroy();
+    setTransitionSystem(stateTransitions: StateTransitionDescription, stateClassStore: any) {
+        this.stateTransitions = stateTransitions;
+        this.stateTransitionSystem = new StateTransitionSystem(this, stateTransitions, stateClassStore);
+        this.stateTransitionSystem.init();
+    }
+
+    setState(state: GameState, reset: boolean = false) {
+        if ( this.currentState && reset ) {
+            PIXI.Loader.shared.removeAllListeners();
+            PIXI.utils.destroyTextureCache();
+            PIXI.Loader.shared.reset();
             this.entityManager.removeAllEntities();
             this.eventQueue.reset();
         }
-        this.start( state );
+        this.currentState = state;
+    }
+
+    start() {
+        if ( !this.currentState ) {
+            throw new Error('There is no current state, try setState(...) or setTransitionSystem(...) before calling start()');
+        }
+
+        if ( !this.running ) {
+            this.running = true;
+            this.currentState.start( this.runGameLoop.bind(this) );
+        }
+    }
+
+    stop(callback: () => void) {
+        const doStop = () => {
+            this.pixiApp.ticker.remove( this.gameLoop, this );
+            if ( this.currentState ) {
+                this.currentState.unstageSystems();
+                this.currentState.unstage();
+            }
+            if ( callback ) {
+                callback();
+            }
+        };
+
+        if ( this.running ) {
+            this.pendingActions.push( doStop );
+        } else {
+            doStop();
+        }
+    }
+
+    private runGameLoop() {
+        this.pixiApp.ticker.add( this.gameLoop, this );
     }
 
     private gameLoop = (delta: number) => {
+        this.running = true;
         const isRenderSystem = function (system: System): system is RenderableSystem {
             return (system as RenderableSystem).render !== undefined;
         }
         this.currentState.getSystems().forEach( s => { s.update(delta); this.eventQueue.dispatchEvents(); } );
         this.currentState.getSystems().forEach( s => isRenderSystem(s) ? (() => { s.render(); this.eventQueue.dispatchEvents(); })() : {} );
         this.currentState.getSystems().forEach( s => s.cleanup() );
-    };
-
-    runGame() {        
-        this.pixiApp.ticker.add( this.gameLoop, this );
-    }
-
-    start(state: GameState) {
-        this.currentState = state;
-        this.currentState.start( this.runGame.bind(this) );
+        this.running = false;
+        while ( this.pendingActions.length > 0 ) {
+            const action = this.pendingActions.shift();
+            action();
+        }
     }
 
 }
